requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
httpx==0.28.1
apscheduler==3.11.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-telegram-bot==21.10
openai==1.61.0
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-mock==3.14.0
